requirements.txt
Django==1.7.1
Pillow==2.6.1
django-ajax-selects==1.3.6
django-extensions==1.5.1
django-selenium==0.9.8
selenium==2.45.0
django-haystack==2.3.1
pysolr==3.3.0
psycopg2==2.6
django-widget-tweaks==1.3
